register controller code and debug

+++ src/utils/cloudinary.js
@@ -1,29 +1,35 @@
-import { v2 as cloudinary } from 'cloudinary'
-import fs from 'fs'
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUDNAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+  secure : true
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try{
-      if(!localFilePath) return null
+  try {
+    if (!localFilePath) return null;
 
-      // upload file on cloudinary
-      const response = await cloudinary.uploader.upload(localFilePath, {    // filepath, options
-        resource_type : 'auto'
-      })
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
 
-      console.log('file is uploaded on cloudinary', response.url)
-      return response
+    // Remove local file after successful upload
+    fs.unlinkSync(localFilePath);
+    return response;
+    
+  } 
+  catch (error) {
 
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-    catch(error){
-      fs.unlinkSync(localFilePath) // The 'unlink' function is used to remove locally saved temporary files in case of failed upload operations
-      return null
-    }
-}
 
-export {uploadOnCloudinary}
+    console.error("Cloudinary Upload Error:", error);
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
